PB-90: Changed Popup share tab to properly update
