+++ src/modules/map/components/LocationPopupPosition.vue
@@ -1,23 +1,13 @@
 <script setup>
 /** Right click pop up which shows the coordinates of the position under the cursor. */
 
-// importing directly the vue component, see https://github.com/ivanvermeyen/vue-collapse-transition/issues/5
-import CollapseTransition from '@ivanv/vue-collapse-transition/src/CollapseTransition.vue'
 import proj4 from 'proj4'
-import tippy from 'tippy.js'
-import { computed, onBeforeUnmount, onMounted, ref, watch } from 'vue'
+import { computed, onMounted, ref, toRefs, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
-import { useRoute } from 'vue-router'
-import { useStore } from 'vuex'
 
 import { requestHeight } from '@/api/height.api'
-import { generateQrCode } from '@/api/qrcode.api'
-import { createShortLink } from '@/api/shortlink.api'
 import { registerWhat3WordsLocation } from '@/api/what3words.api'
-import CesiumPopover from '@/modules/map/components/cesium/CesiumPopover.vue'
-import LocationPopupCopyInput from '@/modules/map/components/LocationPopupCopyInput.vue'
 import LocationPopupCopySlot from '@/modules/map/components/LocationPopupCopySlot.vue'
-import OpenLayersPopover from '@/modules/map/components/openlayers/OpenLayersPopover.vue'
 import {
     LV03Format,
     LV95Format,
@@ -27,37 +17,32 @@ import {
 } from '@/utils/coordinates/coordinateFormat'
 import { WGS84 } from '@/utils/coordinates/coordinateSystems'
 import log from '@/utils/logging'
-import { stringifyQuery } from '@/utils/url-router'
+
+const props = defineProps({
+    coordinate: {
+        type: Boolean,
+        required: true,
+    },
+    clickInfo: {
+        type: Object,
+        required: true,
+    },
+    projection: {
+        type: Object,
+        required: true,
+    },
+    currentLang: {
+        type: Object,
+        required: true,
+    },
+})
+const { coordinate, clickInfo, projection, currentLang } = toRefs(props)
 
 const what3Words = ref(null)
 const height = ref(null)
-const qrCodeImageSrc = ref(null)
-const shareLinkUrlShorten = ref(null)
-const shareLinkUrl = ref(null)
-const showEmbedSharing = ref(false)
 
 const i18n = useI18n()
-const route = useRoute()
-const store = useStore()
 
-const clickInfo = computed(() => store.state.map.clickInfo)
-const currentLang = computed(() => store.state.i18n.lang)
-const projection = computed(() => store.state.position.projection)
-const showIn3d = computed(() => store.state.cesium.active)
-const copyButton = ref(null)
-const copied = ref(false)
-const requestClipboard = ref(false)
-const copyTooltip = ref(null)
-
-const mappingFrameworkSpecificPopup = computed(() => {
-    if (showIn3d.value) {
-        return CesiumPopover
-    }
-    return OpenLayersPopover
-})
-const coordinate = computed(() => {
-    return clickInfo.value?.coordinate
-})
 const coordinateWGS84Metric = computed(() => {
     return proj4(projection.value.epsg, WGS84.epsg, coordinate.value)
 })
@@ -81,76 +66,24 @@ const heightInMeter = computed(() => {
     }
     return ''
 })
-const shareLinkUrlDisplay = computed(() => {
-    return shareLinkUrlShorten.value || shareLinkUrl.value || ''
-})
 
 onMounted(() => {
     if (clickInfo.value) {
         updateWhat3Word()
         updateHeight()
-        if (showEmbedSharing.value) {
-            updateShareLink()
-        }
     }
-})
-
-onMounted(() => {
-    copyTooltip.value = tippy(copyButton.value, {
-        content: i18n.t('copy_success'),
-        arrow: true,
-        placement: 'right',
-        trigger: 'manual',
-        onShow(instance) {
-            setTimeout(() => {
-                instance.hide()
-            }, 1000)
-        },
-    })
-})
-
-onBeforeUnmount(() => {
-    copyTooltip.value.destroy()
 })
 
 watch(clickInfo, (newClickInfo) => {
     if (newClickInfo) {
         updateWhat3Word()
         updateHeight()
-        if (showEmbedSharing.value) {
-            requestClipboard.value = false
-            updateShareLink()
-        }
     }
 })
 watch(currentLang, () => {
     updateWhat3Word()
-    requestClipboard.value = false
-    updateShareLink()
-})
-watch(() => {
-    if (showEmbedSharing.value) {
-        route.query, updateShareLink
-    }
-})
-watch(showEmbedSharing, () => {
-    if (showEmbedSharing.value) {
-        updateShareLink()
-        copyValue()
-    }
-})
-watch(copied, showTooltip)
-watch(shareLinkUrlShorten, () => {
-    if (requestClipboard.value) {
-        copyValue()
-    }
-    requestClipboard.value = false
 })
 
-function clearClick() {
-    store.dispatch('clearClick')
-    showEmbedSharing.value = false
-}
 async function updateWhat3Word() {
     try {
         what3Words.value = await registerWhat3WordsLocation(
@@ -170,55 +103,6 @@ async function updateHeight() {
         log.error(`Failed to get position height`, error)
         height.value = null
     }
-}
-function updateShareLink() {
-    let query = {
-        ...route.query,
-        crosshair: 'marker',
-        center: coordinate.value.join(','),
-    }
-    shareLinkUrl.value = `${location.origin}/#/map?${stringifyQuery(query)}`
-    shortenShareLink(shareLinkUrl.value)
-}
-async function shortenShareLink(url) {
-    try {
-        shareLinkUrlShorten.value = await createShortLink(url)
-        await updateQrCode(shareLinkUrlShorten.value)
-    } catch (error) {
-        log.error(`Failed to shorten Share URL`, error)
-        shareLinkUrlShorten.value = null
-    }
-}
-async function copyValue() {
-    try {
-        copied.value = true
-        if (showEmbedSharing.value) {
-            await navigator.clipboard.writeText(shareLinkUrlShorten.value)
-            copied.value = true
-            setTimeout(() => {
-                copied.value = false
-            }, 1000)
-        }
-    } catch (error) {
-        log.error(`Failed to copy to clipboard:`, error)
-    }
-}
-async function updateQrCode(url) {
-    try {
-        qrCodeImageSrc.value = await generateQrCode(url)
-    } catch (error) {
-        log.error(`Failed to generate qrcode for share url`, error)
-        qrCodeImageSrc.value = null
-    }
-}
-function toggleEmbedSharing() {
-    showEmbedSharing.value = !showEmbedSharing.value
-    if (showEmbedSharing.value) {
-        requestClipboard.value = true
-    }
-}
-function showTooltip() {
-    copyTooltip.value.show()
 }
 </script>
 
@@ -303,23 +187,6 @@ function showTooltip() {
         &-label {
             white-space: nowrap;
         }
-    }
-    &-link {
-        display: flex;
-        align-items: center;
-    }
-    &-qrcode {
-        display: none;
-        text-align: center;
-    }
-    &-coordinates-wgs84-plain {
-        display: inline-block;
-        margin-bottom: 0.1rem;
-    }
-}
-@media (min-height: 0px) {
-    .location-popup-qrcode {
-        display: block;
     }
 }
 </style>
